l'IA peut communiquer via stdin stdout

+++ IA/debugClient.py
@@ -7,14 +7,17 @@ from protocole import *
 
 
 class Debug:
-	def __init__(self):
-		self.p = subprocess.Popen("debug/main.py", stdout=subprocess.PIPE, stdin=subprocess.PIPE)
+	def __init__(self, lvl):
+		if lvl != 0:
+			self.p = subprocess.Popen("debug/main.py", stdout=subprocess.PIPE, stdin=subprocess.PIPE)
 	
 	def log(self, id_cmd, args=None):
-		msg = str(id_cmd)+C_SEP_SEND+str(args)
-		print "debug log : '%s'"%msg
-		self.p.stdin.write(msg+"\n")
+		if lvl != 0:
+			msg = str(id_cmd)+C_SEP_SEND+str(args)
+			print "debug log : '%s'"%msg
+			self.p.stdin.write(msg+"\n")
 	
 	def __del__(self):
-		self.p.terminate()
-		self.p.kill()
+		if lvl != 0:
+			self.p.terminate()
+			self.p.kill()
